chore: move plugins into plugins dir
