+++ packages/gatsby/src/utils/webpack/plugins/__tests__/corejs-resolver.ts
