+++ packages/gatsby/src/utils/webpack/plugins/cache-folder-resolver.ts
