+++ packages/gatsby/src/utils/webpack/plugins/force-css-hmr-for-edge-cases.ts
