+++ packages/gatsby/src/utils/webpack/plugins/webpack-logging.ts
